% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize.R
\name{calcsamplesize}
\alias{calcsamplesize}
\title{Calculates Sample Size}
\usage{
calcsamplesize(
  seuratObj = NULL,
  seuratFile = "./saveddata/mapped_cells.rds",
  dataLocation = "./data/"
)
}
\arguments{
\item{seuratObj}{input seruat object}

\item{seuratFile}{file for seurat rds if \code{seuratObj} is \code{NULL}}

\item{dataLocation}{folder to save sample size data}
}
\value{
a list of \code{data.frame}s
}
\description{
Calculates Sample Size
}
