% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04clustering.R
\name{cluster}
\alias{cluster}
\title{Performs level 0 clustering}
\usage{
cluster(
  seuratObj = NULL,
  seuratFile = "./saveddata/preprocessed_cells.rds",
  resolution = 0.01,
  identities = c(),
  PC = 15,
  umapPC = 20,
  makeplots = TRUE,
  plotLocation = "./images/",
  dotmarkersFile = "./morrismarkers/morris_markers.csv",
  saverds = TRUE,
  write.rdsfile = "./saveddata/clustered_cells.rds"
)
}
\arguments{
\item{seuratObj}{input seruat object}

\item{seuratFile}{file for seurat rds if \code{seuratObj} is \code{NULL}}

\item{resolution}{resolution for clustering}

\item{identities}{cluster renames}

\item{plotLocation}{Folder to put plots if \code{make.plots}}

\item{dotmarkersFile}{csv file with markers for dotplot}

\item{saverds}{saves as an rds if \code{TRUE}.}

\item{write.rdsfile}{path to save rds file to if \code{saverds == TRUE}}

\item{pc}{Number of principal components to use}

\item{umappc}{number of PC to use in UMAP. Morris uses 20}

\item{make.plots}{Make plots?}
}
\value{
a Seurat Object
}
\description{
Performs level 0 clustering
}
