% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/davidplot.R
\name{davidplot}
\alias{davidplot}
\title{Creates dot plot from DAVID results given a specified pathway file}
\usage{
davidplot(
  davidresults = "DAVIDresults/gran chart defaults.txt",
  keptids = "DAVIDresults/gran ids.txt",
  plotname = "./images/david_gran.png",
  height = 4.5
)
}
\arguments{
\item{davidresults}{results text file from DAVID}

\item{keptids}{file (or vector) of ids to plot}

\item{plotname}{name of plot}

\item{height}{height of plot (in inches)}
}
\value{
nothing
}
\description{
Creates dot plot from DAVID results given a specified pathway file
}
