% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07enrichment_analysis.R
\name{enrich}
\alias{enrich}
\title{Performs enrichment analysis}
\usage{
enrich(
  df = NULL,
  DEFolder = "./data/",
  de.q.val = 0.05,
  log2fold.val = 0,
  sign.val = 0,
  level = 0,
  sheetname = "Granulosa",
  savedata = TRUE,
  EAFolder = "./data/"
)
}
\arguments{
\item{df}{input data frame}

\item{DEFolder}{folder for DE data if \code{df} is \code{NULL}}

\item{de.q.val}{maximum adjusted p value for differentially expressed genes}

\item{log2fold.val}{minimum log2fold change needed for DE genes (non-negative)}

\item{sign.val}{-1 for values less than negative. 0 for greater than or less than negative. 1 for values greater}

\item{level}{level to search at (0 or 1) if \code{df} is \code{NULL}}

\item{savedata}{saves as an xslx if \code{TRUE}.}

\item{EAFolder}{path to EA file if \code{savedata == TRUE}}

\item{cell_type}{cell type to search as labelled by name if \code{df} is \code{NULL}}
}
\value{
a list of data.frames
}
\description{
Performs enrichment analysis
}
