% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{makedot}
\alias{makedot}
\title{Creates DotPlots}
\usage{
makedot(
  seuratObj = NULL,
  seuratFile = "./saveddata/mapped_cells.rds",
  plotLocation = "./images/",
  morris.gran.markers = "./morrismarkers/morris_gran_markers.csv",
  morris.mes.merkers = "./morrismarkers/morris_mes_markers.csv",
  nGenes = 5
)
}
\arguments{
\item{seuratObj}{input seruat object}

\item{seuratFile}{file for seurat rds if \code{seuratObj} is \code{NULL}}

\item{plotLocation}{folder to plot images}

\item{morris.gran.markers}{csv where morris top granulosa markers are stored}

\item{morris.mes.merkers}{csv where morris top mesenchyme markers are stored}

\item{nGenes}{number of top genes for each cluster to plot}
}
\value{
nothing
}
\description{
Creates DotPlots
}
