% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05mapping.R
\name{mapcells}
\alias{mapcells}
\title{Maps the cells onto the Morris atlas}
\usage{
mapcells(
  seuratObj = NULL,
  seuratFile = "./saveddata/clustered_cells.rds",
  atlasFile = "ovary_0.rds",
  make.plots = TRUE,
  plotLocation = "./images/",
  atlas.rename = c(),
  umapPC = 20,
  PC = 15,
  anchor.n.trees = 50,
  anchor.k.anchor = 15,
  anchor.k.score = 100,
  marker.folder = "./morrismarkers/",
  saverds = TRUE,
  write.rdsfile = "./saveddata/mapped_cells.rds",
  projumapargs = list()
)
}
\arguments{
\item{seuratObj}{input seruat object}

\item{seuratFile}{file for seurat rds if \code{seuratObj} is \code{NULL}}

\item{atlasFile}{file for Morris atlas}

\item{make.plots}{Make plots?}

\item{plotLocation}{Folder to put plots if \code{make.plots}}

\item{atlas.rename}{.}

\item{umapPC}{number of PC to use in UMAP. Morris uses 20}

\item{PC}{Number of principal components to use}

\item{anchor.n.trees}{\code{n.trees} paramter for \code{FindTransferAnchors}}

\item{anchor.k.anchor}{\code{k.anchor} paramter for \code{FindTransferAnchors}}

\item{anchor.k.score}{\code{k.score} paramter for \code{FindTransferAnchors}}

\item{saverds}{saves as an rds if \code{TRUE}.}

\item{write.rdsfile}{path to save rds file to if \code{saverds == TRUE}}

\item{feature.folder}{folder with \code{.csv} files that contain markers to use}
}
\value{
a Splitted Seurat Object
}
\description{
Maps the cells onto the Morris atlas
}
