% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01mergedata.R
\name{mergedata}
\alias{mergedata}
\title{Grabs 10X data from file, merges into single Seurat object}
\usage{
mergedata(
  min.features = 1000,
  min.cells = 3,
  raw_matrix = "./raw_feature_bc_matrix.h5",
  saverds = TRUE,
  write.rdsfile = "./saveddata/merged_cells.rds"
)
}
\arguments{
\item{min.features}{min.features for \code{CreateSeuratObject} function}

\item{min.cells}{min.cells for \code{CreateSeuratObject} function}

\item{saverds}{saves as an rds if \code{TRUE}.}

\item{write.rdsfile}{path to save rds file to if \code{saverds == TRUE}}
}
\value{
a Seurat Object
}
\description{
Grabs 10X data from file, merges into single Seurat object
}
