% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03preprocessing.R
\name{preprocess}
\alias{preprocess}
\title{Performs Preprocessing}
\usage{
preprocess(
  seuratObj = NULL,
  seuratFile = "./saveddata/filtered_cells.rds",
  make.plots = TRUE,
  plotLocation = "./images/",
  nfeatures = 2000,
  saverds = TRUE,
  write.rdsfile = "./saveddata/preprocessed_cells.rds"
)
}
\arguments{
\item{seuratObj}{input seruat object}

\item{seuratFile}{file for seurat rds if \code{seuratObj} is \code{NULL}}

\item{make.plots}{Make plots?}

\item{plotLocation}{Folder to put plots if \code{make.plots}}

\item{nfeatures}{features to select}

\item{saverds}{saves as an rds if \code{TRUE}.}

\item{write.rdsfile}{path to save rds file to if \code{saverds == TRUE}}
}
\value{
a Seurat Object
}
\description{
Performs Preprocessing
}
