% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02QC.R
\name{QC}
\alias{QC}
\title{Performs Quality Control}
\usage{
QC(
  seuratObj = NULL,
  seuratFile = "./saveddata/merged_cells.rds",
  make.plots = TRUE,
  plotLocation = "./images/",
  nFeature.lower = 3500,
  nFeature.upper = 7000,
  percent.mt.upper = 5,
  nCount.rna.upper = 60000,
  saverds = TRUE,
  write.rdsfile = "./saveddata/filtered_cells.rds"
)
}
\arguments{
\item{seuratObj}{input seruat object}

\item{seuratFile}{file for seurat rds if \code{seuratObj} is \code{NULL}}

\item{make.plots}{Make plots?}

\item{plotLocation}{Folder to put plots if \code{make.plots}}

\item{nFeature.lower}{removes cells with \code{nFeature < nFeature.lower}}

\item{nFeature.upper}{removes cells with \code{nFeature > nFeature.upper}}

\item{percent.mt.upper}{removes cells with \code{percemt_mt > percent.mt.upper}}

\item{nCount.rna.upper}{removes cells with \code{nCount_RNA > nCount.rna.upper}}

\item{saverds}{saves as an rds if \code{TRUE}.}

\item{write.rdsfile}{path to save rds file to if \code{saverds == TRUE}}
}
\value{
a Seurat Object
}
\description{
Performs Quality Control
}
