% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulationplots.R
\name{regulationplots}
\alias{regulationplots}
\title{creates plots of up/downreugulated counts}
\usage{
regulationplots(DEFile1, DEFile2, order1, order2, padding, out1, out2)
}
\arguments{
\item{DEFile1}{DE level 0}

\item{DEFile2}{DE level 1}

\item{order1}{celltype order level 0}

\item{order2}{celltype order level 1}

\item{padding}{horizontal padding in graphs}

\item{out1}{output image level 0}

\item{out2}{output image level 1}
}
\value{
nothing
}
\description{
creates plots of up/downreugulated counts
}
