% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06differential_expression.R
\name{calcDE}
\alias{calcDE}
\title{Performs differential expression}
\usage{
calcDE(
  seuratObj = NULL,
  seuratFile = "./saveddata/mapped_cells.rds",
  logfc.threshhold = 0.1,
  savedata = TRUE,
  DEFolder = "./data/"
)
}
\arguments{
\item{seuratObj}{input seruat object}

\item{seuratFile}{file for seurat rds if \code{seuratObj} is \code{NULL}}

\item{logfc.threshhold}{minimum logfc to be considered for DE.}

\item{savedata}{saves as an xslx if \code{TRUE}.}

\item{DEFolder}{path to DE file if `savedata == TRUE}
}
\value{
list of lists of \code{data.frame}s \link{level}
}
\description{
Performs differential expression
}
